requirements.txt
Django==4.1.3
psycopg2-binary==2.9.5
python-telegram-bot==13.14